Idempotent provision

+++ Vagrantfile
@@ -73,7 +73,12 @@ Vagrant.configure("2") do |config|
   # documentation for more information about their specific syntax and use.
   config.vm.provision "shell", inline: <<-SHELL
     cd uwece651f16_js
-    ln -s /home/vagrant/node_modules node_modules
-    ln -s /home/vagrant/bower_components client/bower_components
+    if [ ! -e 'node_modules']; then
+      ln -s /home/vagrant/node_modules node_modules
+    fi
+
+    if [ ! -e 'client/bower_components']; then
+      ln -s /home/vagrant/bower_components client/bower_components
+    fi
   SHELL
 end
